fix models

+++ HealthyNutritionApp.Application/Models/NutritionFact.cs
@@ -2,9 +2,9 @@
 {
     public class NutritionFact
     {
-        public int Calories { get; set; }
-        public int Protein { get; set; }
-        public int Carbs { get; set; }
-        public int Fat { get; set; }
+        public double Calories { get; set; }
+        public double Protein { get; set; }
+        public double Carbs { get; set; }
+        public double Fat { get; set; }
     }
 }
